pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PathwayOracle"
version = "2.0.1"
dependencies = [
    "langchain-community",
    "langchain-openai",
    "langchain",
    "langchain-core",
    "langchain-text-splitters",
    "python-dotenv",
    "scikit-learn",
    "hdbscan",
    "numpy",
    "matplotlib",
    "requests",
    "neo4j",
    "torch",
    "torchvision",
    "torchaudio",
    "transformers",
]
authors = [
  { name="Yahya Al Sabeh", email="yahyaalsabeh@gmail.com" },
]
description = "An LLM-empowered, KG-Driven, Gene Pathway Analysis Tool"
readme = "README.md"
requires-python = "==3.10.15"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/yoyo4581/PathwayOracle"
Issues = "https://github.com/yoyo4581/PathwayOracle/issues"